refactor(shared): split pure diff and timeline cores from the file-based runners
